+++ app/src/organisms/Desktop/Devices/RobotSettings/RobotSettingsComplianceReady/__tests__/complianceReadySettingsHelper.test.ts
@@ -1,9 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import {
-  getAuthPatchForInputChange,
-  resolveComplianceReadyToggleChange,
-} from '../complianceReadySettingsHelper'
+import { getAuthPatchForInputChange } from '../complianceReadySettingsHelper'
 
 import type { FieldValues } from '../complianceReadySettingsTypes'
 
@@ -96,124 +93,5 @@ describe('getAuthPatchForInputChange', () => {
         'passwordComplexityEnabled'
       )
     ).toEqual({ data: { passwordComplexityMinimumLength: 12 } })
-  })
-})
-
-describe('resolveComplianceReadyToggleChange', () => {
-  it('should not patch when enabling a UI-only parent', () => {
-    expect(
-      resolveComplianceReadyToggleChange(
-        'passwordResetEnabled',
-        BASE_FIELD_VALUES,
-        { childFieldIds: ['passwordResetTime'] }
-      )
-    ).toEqual({})
-  })
-
-  it('should patch null for auth children when disabling a UI-only parent', () => {
-    expect(
-      resolveComplianceReadyToggleChange(
-        'passwordResetEnabled',
-        {
-          ...BASE_FIELD_VALUES,
-          passwordResetEnabled: true,
-          passwordResetTime: '30',
-        },
-        { childFieldIds: ['passwordResetTime'] }
-      )
-    ).toEqual({
-      authPatch: { data: { passwordResetTime: null } },
-    })
-  })
-
-  it('should null all nested auth fields when disabling UI-only parent with multiple children', () => {
-    expect(
-      resolveComplianceReadyToggleChange(
-        'passwordComplexityEnabled',
-        {
-          ...BASE_FIELD_VALUES,
-          passwordComplexityEnabled: true,
-          passwordComplexitySpecialCharacters: true,
-          passwordComplexityMinimumLength: '8',
-        },
-        {
-          childFieldIds: [
-            'passwordComplexitySpecialCharacters',
-            'passwordComplexityMinimumLength',
-          ],
-        }
-      )
-    ).toEqual({
-      authPatch: {
-        data: {
-          passwordComplexitySpecialCharacters: null,
-          passwordComplexityMinimumLength: null,
-        },
-      },
-    })
-  })
-
-  it('should patch auth server when toggling a child under an enabled UI-only parent', () => {
-    expect(
-      resolveComplianceReadyToggleChange(
-        'passwordComplexitySpecialCharacters',
-        {
-          ...BASE_FIELD_VALUES,
-          passwordComplexityEnabled: true,
-          passwordComplexitySpecialCharacters: false,
-          passwordComplexityMinimumLength: '8',
-        },
-        { parentFieldId: 'passwordComplexityEnabled' }
-      )
-    ).toEqual({
-      authPatch: { data: { passwordComplexitySpecialCharacters: true } },
-    })
-  })
-
-  it('should not patch when toggling a child with UI-only parent off', () => {
-    expect(
-      resolveComplianceReadyToggleChange(
-        'passwordComplexitySpecialCharacters',
-        BASE_FIELD_VALUES,
-        { parentFieldId: 'passwordComplexityEnabled' }
-      )
-    ).toEqual({})
-  })
-
-  it('should patch auth server when toggling a standalone auth setting', () => {
-    expect(
-      resolveComplianceReadyToggleChange(
-        'requireAdminCredsWhenUpdatingRobotSoftware',
-        BASE_FIELD_VALUES
-      )
-    ).toEqual({
-      authPatch: {
-        data: { requireAdminCredsWhenUpdatingRobotSoftware: false },
-      },
-    })
-  })
-
-  it('should patch robot server when toggling a robot server setting', () => {
-    expect(
-      resolveComplianceReadyToggleChange(
-        'requireSignoffForProtocolLog',
-        BASE_FIELD_VALUES
-      )
-    ).toEqual({
-      robotServerAccessControlPatch: {
-        data: { requireSignoffForProtocolLog: false },
-      },
-    })
-  })
-
-  it('should patch auth server when toggling a real auth parent with children', () => {
-    expect(
-      resolveComplianceReadyToggleChange(
-        'requireReasonForInteraction',
-        BASE_FIELD_VALUES
-      )
-    ).toEqual({
-      authPatch: { data: { requireReasonForInteraction: false } },
-    })
   })
 })
